Add lap timer functionality to RacerViewModel
This commit adds the ability to start, stop and catch lap times in the RacerViewModel. It also includes a DispatcherTimer that updates the LapTimer property every 100 milliseconds. The LapTimer is reset when a lap time is caught and added to the Laps collection. A Finished property has been added to indicate if the racer has finished their laps. The StartTimerCommand, StopTimerCommand, and CatchLapTimeCommand have been implemented along with their respective CanExecute methods. Finally, an event handler for RacerLapsCollectionChanged was added to update RaceProfit whenever a racer's laps are changed.