Add lap distance to race and racer models
This commit adds a new property, LapDistanceMiles, to the RaceModel and RacerLapModel classes. The value of this property is used to calculate the lap speed in miles per hour. Additionally, the Bootstrapper class now uses a different delegate signature for creating instances of RacerViewModel. Finally, the RacesSummaryViewModel class now sets the LapDistanceMiles property when creating an instance of RaceModel.